Feat/perfis

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -2,24 +2,30 @@ package br.com.forum_hub.domain.resposta;
 
 import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
+import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
+import br.com.forum_hub.infra.security.HierarquiaService;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class RespostaService {
-    private final RespostaRepository repository;
-    private final TopicoService topicoService;
 
-    public RespostaService(RespostaRepository repository, TopicoService topicoService) {
-        this.repository = repository;
-        this.topicoService = topicoService;
-    }
+    @Autowired
+    private RespostaRepository repository;
+
+    @Autowired
+    private TopicoService topicoService;
+
+    @Autowired
+    private HierarquiaService hierarquiaService;
 
     @Transactional
-    public Resposta cadastrar(DadosCadastroResposta dados, Long idTopico) {
+    public Resposta cadastrar(DadosCadastroResposta dados, Long idTopico, Usuario autor) {
         var topico = topicoService.buscarPeloId(idTopico);
 
         if(!topico.estaAberto()) {
@@ -32,13 +38,18 @@ public class RespostaService {
 
         topico.incrementarRespostas();
 
-        var resposta = new Resposta(dados, topico);
+        var resposta = new Resposta(dados, topico, autor);
         return repository.save(resposta);
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getTopico().getAutor(), "ROLE_MODERADOR")){
+            throw new AccessDeniedException("Você não pode editar essa resposta!");
+        }
+
         return resposta.atualizarInformacoes(dados);
     }
 
@@ -47,11 +58,16 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
 
         var topico = resposta.getTopico();
-        if(topico.getStatus() == Status.RESOLVIDO)
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR")) {
+            throw new AccessDeniedException("Você não pode marcar essa resposta como solução!");
+        }
+
+        if (topico.getStatus() == Status.RESOLVIDO)
             throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
 
         topico.alterarStatus(Status.RESOLVIDO);
@@ -59,17 +75,21 @@ public class RespostaService {
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR")){
+            throw new AccessDeniedException("Você não pode excluir essa resposta!");
+        }
 
         repository.deleteById(id);
 
         topico.decrementarRespostas();
         if (topico.getQuantidadeRespostas() == 0)
-            topico.alterarStatus(Status.NAO_RESPONDIDO);
+        topico.alterarStatus(Status.NAO_RESPONDIDO);
         else if(resposta.ehSolucao())
-            topico.alterarStatus(Status.RESPONDIDO);
+        topico.alterarStatus(Status.RESPONDIDO);
     }
 
     public Resposta buscarPeloId(Long id) {
